Fix AndroidLogger Move utility methods and classes to separate file
