+++ leveldb/build.gradle
@@ -102,6 +102,12 @@ android {
         testHandleProfiling true
         testFunctionalTest true
 
+        externalNativeBuild {
+            cmake {
+                cppFlags "-fno-rtti"
+            }
+        }
+
         ndk {
             abiFilters 'armeabi-v7a', 'arm64-v8a', 'x86', 'x86_64'
             moduleName "leveldb"
